DEP-207 습관보관함 페이지 UI 적용 (#78)
* ui: 보관된 습관 목록 ui 적용

# Conflicts:
#	presentation/mypage/src/main/res/layout/activity_archived_habit.xml
#	presentation/mypage/src/main/res/values/strings.xml

* ui: 습관 보관함 삭제하기 ui 적용

# Conflicts:
#	data/src/main/java/com/depromeet/threedays/data/datasource/habit/HabitRemoteDataSourceImpl.kt
#	presentation/mypage/src/main/res/values/strings.xml

* refactor: adapter 에서 viewModel 사용하지 않도록 callback 전달

* refactor: 불필요한 method 제거

* refactor: 투명도 설정 변경 (코드 -> xml)

+++ data/src/main/java/com/depromeet/threedays/data/datasource/habit/HabitRemoteDataSource.kt
@@ -9,5 +9,5 @@ interface HabitRemoteDataSource {
     suspend fun getArchivedHabits(): List<HabitEntity>
     //fun getHabit(habitId: Long): HabitEntity
     suspend fun updateHabit(habitId: Long)
-    suspend fun deleteHabit(habitId: Int): Any
+    suspend fun deleteHabit(habitId: Long): Any
 }
