+++ data/src/main/java/com/depromeet/threedays/data/repository/HabitRepositoryImpl.kt
@@ -55,7 +55,7 @@ class HabitRepositoryImpl @Inject constructor(
 
     }
 
-    override suspend fun deleteHabit(habitId: Int): Flow<DataState<Any>> =
+    override suspend fun deleteHabit(habitId: Long): Flow<DataState<Any>> =
         flow {
             emit(DataState.loading())
             val response = habitRemoteDataSource.deleteHabit(habitId)
